Modify test script

+++ test/unit/test.sh
@@ -16,6 +16,7 @@ if [ ! -d "$working_dir1" ]; then
   exit
 fi
 cd $working_dir1
+rm $working_dir1/gtestlog.txt
 
 #Gtest functions
 unittest="$working_dir1/hcfft_Create_destroy_Plan"
@@ -24,13 +25,17 @@ runcmd1="$unittest >> gtestlog.txt"
 eval $runcmd1
 
 Log_file="$working_dir1/gtestlog.txt"
-if grep -q FAILED "$Log_file";
-then
+if [ ! -s "$Log_file" ]; then
+  echo "${red}GTEST IS NOT WORKING....${reset}"
+else
+  if grep -q FAILED "$Log_file";
+  then
     echo "${red}hcfft_Create_destroy_Plan               ----- [ FAILED ]${reset}"
   elif grep -q PASSED "$Log_file";
   then
     echo "${green}hcfft_Create_destroy_Plan             ----- [ PASSED ]${reset}"
     rm $working_dir1/gtestlog.txt
+  fi
 fi
 
 testdirectories=(hcfft_1D_transform hcfft_2D_transform hcfft_3D_transform)
@@ -40,6 +45,7 @@ for i in 0 1 2
 do
   working_dir1="$current_work_dir/../../build/test/unit/gtest/${testdirectories[$i]}/bin/"
   cd $working_dir1
+  rm $working_dir1/gtestlog.txt
 
   #Gtest functions
   unittest="$working_dir1/${testdirectories[$i]}"
@@ -48,13 +54,17 @@ do
   eval $runcmd1
 
   Log_file="$working_dir1/gtestlog.txt"
-  if grep -q FAILED "$Log_file";
-  then
+  if [ ! -s "$Log_file" ]; then
+    echo "${red}GTEST IS NOT WORKING....${reset}"
+  else
+    if grep -q FAILED "$Log_file";
+    then
       echo "${red}${testdirectories[$i]}               ----- [ FAILED ]${reset}"
-  elif grep -q PASSED "$Log_file";
-  then
+    elif grep -q PASSED "$Log_file";
+    then
       echo "${green}${testdirectories[$i]}             ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
+    fi
   fi
 
   #Gtest functions
@@ -64,13 +74,17 @@ do
   eval $runcmd1
 
   Log_file="$working_dir1/gtestlog.txt"
-  if grep -q FAILED "$Log_file";
-  then
-      echo "${red}${testdirectories[$i]}_double               ----- [ FAILED ]${reset}"
-  elif grep -q PASSED "$Log_file";
-  then
+  if [ ! -s "$Log_file" ]; then
+    echo "${red}GTEST IS NOT WORKING....${reset}"
+  else
+    if grep -q FAILED "$Log_file";
+    then
+      echo "${red}${testdirectories[$i]}_double             ----- [ FAILED ]${reset}"
+    elif grep -q PASSED "$Log_file";
+    then
       echo "${green}${testdirectories[$i]}_double           ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
+    fi
   fi
 
   #Gtest functions
@@ -80,12 +94,16 @@ do
   eval $runcmd1
 
   Log_file="$working_dir1/gtestlog.txt"
-  if grep -q FAILED "$Log_file";
-  then
-      echo "${red}${testdirectories[$i]}_padding               ----- [ FAILED ]${reset}"
-  elif grep -q PASSED "$Log_file";
-  then
+  if [ ! -s "$Log_file" ]; then
+    echo "${red}GTEST IS NOT WORKING....${reset}"
+  else
+    if grep -q FAILED "$Log_file";
+    then
+      echo "${red}${testdirectories[$i]}_padding             ----- [ FAILED ]${reset}"
+    elif grep -q PASSED "$Log_file";
+    then
       echo "${green}${testdirectories[$i]}_padding           ----- [ PASSED ]${reset}"
       rm $working_dir1/gtestlog.txt
+    fi
   fi
 done
